+++ src/index.ts
@@ -8,6 +8,7 @@ import StatusChip from './components/StatusChip';
 import Input from './components/Input';
 import LoadingFullScreen from './components/LoadingFullScreen';
 import Icon from './components/Icon';
+import Select from './components/Select';
 import {palette} from './theme/palette';
 import RadioButton from './components/RadioButton';
 
@@ -24,4 +25,5 @@ export {
 	palette,
 	LoadingFullScreen,
 	RadioButton,
+	Select,
 };